Final Commit
Added a video clip into the aside with controls for the user to use. Created a .vtt file containing captions for the video, which was then linked by using the track element, and stylized the captions by changing the font size and the color. Added a media query for devices with minimum widths of 521px to float the video to the right. Ready for review.